+++ lib/ecg/real_time/real_time.dart
@@ -1,8 +1,9 @@
 import "package:flutter/material.dart";
+import "package:flutter_reactive_ble/flutter_reactive_ble.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:functional_widget_annotation/functional_widget_annotation.dart";
 
-import "../../device_manager/device.dart";
+import "../../device/device.dart";
 import "chart.dart";
 import "device_not_available.dart";
 import "heart_rate/heart_rate.dart";
@@ -11,7 +12,8 @@ part "real_time.g.dart";
 
 @cwidget
 Widget _realTime(BuildContext context, WidgetRef ref) {
-  final deviceAvailable = ref.watch(connectedProvider).value ?? true;
+  final deviceAvailable = ref.watch(connectionStateProvider).value !=
+      DeviceConnectionState.disconnected;
   if (!deviceAvailable) {
     return const DeviceNotAvailable();
   }
@@ -19,8 +21,8 @@ Widget _realTime(BuildContext context, WidgetRef ref) {
   final isPortrait = MediaQuery.of(context).orientation == Orientation.portrait;
 
   if (isPortrait) {
-    return Column(
-      children: const [
+    return const Column(
+      children: [
         Expanded(child: HeartRateWidget()),
         Expanded(flex: 5, child: RealTimeChart()),
       ],