+++ lib/utils/database.dart
@@ -1,12 +1,13 @@
 import "package:flutter/material.dart";
 import "package:isar/isar.dart";
 import "package:logging/logging.dart";
+import "package:path_provider/path_provider.dart";
 import "package:quiver/time.dart";
 import "package:shared_preferences/shared_preferences.dart";
 
 import "../analytics/data_types.dart";
-import "../device_manager/device.dart";
-import "../device_manager/fake_device.dart";
+import "../device/device.dart";
+import "../device/fake_device.dart";
 
 part "database.g.dart";
 
@@ -240,11 +241,15 @@ late final Isar _isar;
 
 Future<void> initDatabase() async {
   prefs = await SharedPreferences.getInstance();
-  _isar = await Isar.open([
-    BeatSchema,
-    SamplePointSchema,
-    FakeSamplePointSchema,
-  ]);
+  final dir = await getApplicationDocumentsDirectory();
+  _isar = await Isar.open(
+    directory: dir.path,
+    [
+      BeatSchema,
+      SamplePointSchema,
+      FakeSamplePointSchema,
+    ],
+  );
 }
 
 @visibleForTesting